fix: split server and runtime code in plugin-data-loader (#3224)
* fix: split server and runtime code in plugin-data-loader

* docs: add changeset

* chore: update lock file
