+++ .changeset/ten-shrimps-sell.md
@@ -1,0 +1,6 @@
+---
+'@modern-js/plugin-data-loader': patch
+---
+
+fix: split server and runtime code in plugin-data-loader
+fix: 分离 plugin-data-loader 中服务端和客户端运行时的代码
