+++ packages/cli/plugin-data-loader/src/server/index.ts
@@ -4,8 +4,7 @@ import fs from 'fs';
 import type { ServerPlugin } from '@modern-js/server-core';
 import type { ServerRoute } from '@modern-js/types';
 import { MAIN_ENTRY_NAME, SERVER_BUNDLE_DIRECTORY } from '@modern-js/utils';
-
-import { matchEntry, ServerContext } from '../runtime';
+import { matchEntry, ServerContext } from '../common/utils';
 
 export default (): ServerPlugin => ({
   name: '@modern-js/plugin-data-loader',
